Ref
Subdivision of main.ts file
